Update tests structure to add parser tests
