+++ package.json
@@ -13,7 +13,7 @@
     "main": "src/tester.js",
     "scripts": {
         "test": "node src/tester.js --test-folder tests/scripts/ --failure-folder failures --doc-path tests/html_files/",
-        "api-test": "node tests/test-js/main.js",
+        "api-test": "node tests/test-js/api.js",
         "lint": "eslint src tests/test-js",
         "lint-fix": "eslint --fix src tests/test-js"
     },
